updated code with main

+++ app/(groupControl)/invoiceControl/page.jsx
@@ -35,6 +35,7 @@ import {
   fetchDataAPI,
   disableEdit,
   disableAdd,
+  disablePrint,
 } from "@/services/auth/FormControl.services.js";
 import {
   viewIcon,
@@ -98,7 +99,7 @@ import PrintModal from "@/components/Modal/printModal.jsx";
 
 export default function StickyHeadTable() {
   const router = useRouter();
-   const { clientId } = getUserDetails();
+  const { clientId } = getUserDetails();
   const searchParams = useSearchParams();
   const search = JSON.parse(searchParams.get("menuName")).id;
   const [menuSearch, setMenuSearch] = useState("");
@@ -273,15 +274,18 @@ export default function StickyHeadTable() {
   };
 
   const validateEdit = async (tableName, recordId) => {
+    //alert("working")
     const requestBody = {
       tableName: tableName,
       recordId: recordId.id,
+      clientId: clientId,
     };
     const data = await disableEdit(requestBody);
-    if (data.success === true) {
+    if (data.success === false) {
       setParaText(data.message);
       //setIsError(true);
       setOpenModal((prev) => !prev);
+      return;
     } else {
       addEditController(recordId);
     }
@@ -641,7 +645,7 @@ export default function StickyHeadTable() {
       setOpenModal((prev) => !prev);
     }
     // API call for delete
-    if (conformData.value) {
+    if (conformData.value && deleteData) {
       const payloadData = { ...deleteData };
       try {
         const responseData = await deleteMasterRecord(payloadData);
@@ -658,6 +662,8 @@ export default function StickyHeadTable() {
         console.error("Error in API call:", error);
         toast.error("An error occurred while processing the request");
       }
+    } else {
+      setOpenModal((prev) => !prev);
     }
   };
 
@@ -766,10 +772,15 @@ export default function StickyHeadTable() {
   const handleCustomRowsPerPageChange = (event) => {
     const value = parseInt(event.target.value, 10);
     if (!isNaN(value) && value >= 0) {
+      sessionStorage?.setItem("rowsPerPage", value);
       setRowsPerPage(value);
       setSelectedPageNumber(page);
     }
   };
+  useEffect(() => {
+    const storedRowsPerPage = sessionStorage.getItem("rowsPerPage");
+    setRowsPerPage(storedRowsPerPage ? parseInt(storedRowsPerPage) : 17);
+  }, [sessionStorage.getItem("rowsPerPage"), search]);
 
   function pageSelected(selectedValue) {
     // console.log("pageSelected - - - - - ", selectedValue);
@@ -779,23 +790,22 @@ export default function StickyHeadTable() {
   }
 
   useEffect(() => {
-      // console.log('checkMenuId', sessionStorage.getItem('menuId') == search);
-      if (sessionStorage?.getItem('menuId') == search) {
-        const menuId = JSON.parse(sessionStorage.getItem('advanceSearch'));
-        const dynamic = JSON.parse(sessionStorage.getItem('dynamic'));
-        console.log("dynamic", dynamic)
-        setadvanceSearch(menuId);
-        setDynamic(dynamic);
-        // setIsAdvanceSearchOpen(true);
-      }
-      else {
-        sessionStorage.removeItem('advanceSearch');
-        sessionStorage.removeItem('menuId');
-        sessionStorage.removeItem('dynamic');
-        setadvanceSearch({});
-        setDynamic([]);
-      }
-    }, [search])
+    // console.log('checkMenuId', sessionStorage.getItem('menuId') == search);
+    if (sessionStorage?.getItem("menuId") == search) {
+      const menuId = JSON.parse(sessionStorage.getItem("advanceSearch"));
+      const dynamic = JSON.parse(sessionStorage.getItem("dynamic"));
+      console.log("dynamic", dynamic);
+      setadvanceSearch(menuId);
+      setDynamic(dynamic);
+      // setIsAdvanceSearchOpen(true);
+    } else {
+      sessionStorage.removeItem("advanceSearch");
+      sessionStorage.removeItem("menuId");
+      sessionStorage.removeItem("dynamic");
+      setadvanceSearch({});
+      setDynamic([]);
+    }
+  }, [search]);
   // Function to handle sorting when a column header is clicked
   const handleSortBy = (elem) => {
     // If the same column is clicked again, toggle the sorting order
@@ -1116,9 +1126,9 @@ export default function StickyHeadTable() {
       Object.assign(tempObj, d.advanceSearch);
     });
     setadvanceSearch(tempObj);
-    sessionStorage.setItem('advanceSearch', JSON.stringify(tempObj));
-    sessionStorage.setItem('menuId', search);
-    sessionStorage.setItem('dynamic', JSON.stringify(dynamic));
+    sessionStorage.setItem("advanceSearch", JSON.stringify(tempObj));
+    sessionStorage.setItem("menuId", search);
+    sessionStorage.setItem("dynamic", JSON.stringify(dynamic));
     setSearchOpen(false);
   };
 
@@ -2193,6 +2203,7 @@ export default function StickyHeadTable() {
                                 onClick={async () => {
                                   handlePrint(row);
                                   setModalVisible(true);
+                                  //validateEditPrint(tableName, row)
                                 }}
                               />
                               {/* )} */}
@@ -2235,6 +2246,7 @@ export default function StickyHeadTable() {
                     submittedRecordId={submittedRecordId}
                     submittedMenuId={submittedMenuId}
                     openPrintModal={openPrintModal}
+                    tableName={tableName}
                     pageType={"searchPage"}
                   />
                 )}
